requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
dateparser==1.2.2
discord-webhook==1.4.1
discord.py==2.5.2
distro==1.9.0
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
loguru==0.7.3
multidict==6.6.3
numpy==2.3.1
openai==1.97.0
pandas==2.3.1
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
python-binance==1.0.29
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
ta==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
